+++ docs/CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Changed
+-------
+- Retry data transfer if ``botocore.exceptions.ClientError`` is raised during
+  transfer
+
+
 ===================
 28.1.0 - 2021-05-17
 ===================